+++ index.html
@@ -231,7 +231,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script type="module" src="./js/functions.js"></script>
   <script type="module" src="./js/main.js"></script>
 </body>
 </html>
